add method in stub and rename fields
