+++ engine/node/score-node-impl/src/main/java/io/cloudslang/engine/node/services/StubQueueConfigurationDataServiceImpl.java
@@ -23,4 +23,19 @@ public class StubQueueConfigurationDataServiceImpl implements QueueConfiguration
 	public QueueDetails getQueueConfigurations() {
 		return null;
 	}
+
+	@Override
+	public boolean testQueueConnection(QueueDetails queueDetailsVO) {
+		return false;
+	}
+
+	@Override
+	public QueueDetails retrieveQueueConfiguration() {
+		return null;
+	}
+
+	@Override
+	public void configureQueueParameters(QueueDetails queueDetailsVO) {
+
+	}
 }
